VERSION 0.7.0
Added -patch mode because I needed it.

This time I had a drive which got unreadable sectors.
Of course this happend friday night, so it needs days for a replacement drive.

The drive was part of a ZFS raidz3 pool and S.M.A.R.T. was unwilling to remap
the defective sectors, so I tried this approach:

While the drive was still used for ZFS I applied -patch mode to it.
Afterwards I only had to scrub ZFS again to repair the damage.

The problems did not vanish completely, but it is much better now,
so I borrowed some time until I find the time to replace the drive.

Note: I do not recommend this procedure to anybody.
The risk to loose data is far too high if you do not exactly what you are doing.
Hoever, ZFS is like a cockroach, it's die hard.

------------------------------------------------------------------------

VERSION 0.6.4

This fixes a bug in a library part, from which it is unknown if diskus is affected.

------------------------------------------------------------------------

VERSION 0.6.3

CygWin Bugfixes:  Now compiles under CygWin.

------------------------------------------------------------------------

VERSION 0.6.2

Bugfixes:
Option -xd now works

Previously Option -x always was active, this was not intended.

------------------------------------------------------------------------

VERSION 0.6.1

Bugfixes:
Option -null now works again.
Option -jump (experimental feature) now works for small blocksizes.

There still is no way to use -jump and -freshen such,
that defective sectors are overwritten,
to give the drive a chance to remap the sector.

------------------------------------------------------------------------

VERSION 0.6.0

Option -xd added
check mode changed
dump mode now uses extended xd from tinolib

Warning:  Install path now is /usr/local/bin/
so be sure to rm old version in $HOME/bin/

------------------------------------------------------------------------

VERSION 0.5.1

Output changed (using my scale library)
Also some small changes to make read mode more reliable.

------------------------------------------------------------------------

VERSION 0.5.0

-freshen mode and -to option added

This new feature is again somehwat experimental.  With -freshen, all
data read is rewritten to disk.  It is done overlapping, such that each
sector is read and written two times.  The idea is to "freshen" the magnetic
data by overwriting it with itself.  Before you ask:  I needed this to
have some drives remap weakened (eroded) sectors which took ages to read.

The -to option is to give an end option for operation.  This way you can
give an end-position (positive value) or delta (number of bytes to operate
on with a negative value).  For example this will read a drive from
200MiB to 400MiB:

diskus -start 200M -to -200M -read /dev/sda

------------------------------------------------------------------------

VERSION 0.4.1

Error reporting changed to be better readable, some library changes.

Nearly untested.

------------------------------------------------------------------------

VERSION 0.4.0

Option -jump added (because I needed it)

This feature is very experimental and only works for read modes currently.
It tries to jump over errors by skipping to the next 4096 byte boundary.
It has not been tested thoroughly and if run with the -async option
(which automatically might kick in in several cases) this will behave
a little bit unreliably.

------------------------------------------------------------------------

VERSION 0.3.4

More suffixes: SC and PEZY

------------------------------------------------------------------------

VERSION 0.3.3

IO buffer now is aligned to a page boundary,
such that the -async option is no more needed
in most times.

------------------------------------------------------------------------

VERSION 0.3.2

-write mode now no more valid for read modes

I always mix this myself, so added a fatal error for this.

------------------------------------------------------------------------

VERSION 0.3.1

Reporting corrected

The result printed with -gen mode differed from that with -check mode,
which was really puzzling.  Now both shall show the same counts.

(Previously the read-loop showed the correct count, while the
write-loop showed the count which would be reached if not EOF
interfered, which was too high.)

------------------------------------------------------------------------

VERSION 0.3.0

Bugfix release, better reporting and return value,
Option -quiet to suppress progress status

Diskus previously was not very detailed in the output,
it now prints more information if -quiet is not present.

The "dump" function 0.2.x did no more work correctly if option -expand
was not present, this is fixed now.

------------------------------------------------------------------------

VERSION 0.2.0

New option -expand

Because I needed it

------------------------------------------------------------------------

VERSION 0.1.1

Removed bug in -dump.

On characters above DEL -dump was unable to detect repeated sequences.

------------------------------------------------------------------------

VERSION 0.1.0

Removed minor bugs.

There was a bug when reading devices without option -async present.
Newer kernels apparently do not allow to do that.  The workaround was
to add the -async option.  This is now automatically applied.

Also if the start value was not a multiple of the sector size strange
things happened.  The start value now is rounded.

------------------------------------------------------------------------

VERSION 0.0.1

Looking at the sources I already found the first bug.

Wasn't that quick?

------------------------------------------------------------------------

VERSION 0.0.0

Just a first version.

There is a bug in that it tries to read past the device end.
In this case you do not get EOF, instead you see an IO error
and see some ugly kernel entries in the syslog.